adding exception handler

+++ src/main/java/com/example/taskmanager/controller/TaskController.java
@@ -2,12 +2,16 @@ package com.example.taskmanager.controller;
 
 import com.example.taskmanager.dto.TaskCreateDto;
 import com.example.taskmanager.dto.TaskDto;
+import com.example.taskmanager.exception.TaskNotFoundException;
 import com.example.taskmanager.model.Task;
 import com.example.taskmanager.service.TaskService;
 import jakarta.validation.Valid;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.Objects;
 
 @RestController
 @RequestMapping(path = "/api/tasks")
@@ -20,33 +24,49 @@ public class TaskController {
     }
 
     @PostMapping
-    public TaskDto createTask(@RequestBody @Valid TaskCreateDto taskCreateDto) {
-        Task task =  taskService.createTask(taskCreateDto);
-        return taskService.convertEntityToDto(task);
+    public ResponseEntity<TaskDto> createTask(@RequestBody @Valid TaskCreateDto taskCreateDto) {
+        Task task = taskService.createTask(taskCreateDto);
+        TaskDto taskDto = taskService.convertEntityToDto(task);
+
+        return ResponseEntity.status(HttpStatus.CREATED).body(taskDto);
     }
 
     @PutMapping(path = "{id}")
-    public TaskDto updateTask(@PathVariable(name = "id")  Long id, @RequestBody TaskCreateDto taskCreateDto) {
+    public ResponseEntity<TaskDto> updateTask(@PathVariable(name = "id")  Long id, @RequestBody TaskCreateDto taskCreateDto) {
         Task task =  taskService.findTaskById(id);
+        if(Objects.isNull(task)){
+            throw new TaskNotFoundException("Task with ID " + id + " not found");
+        }
         task = taskService.updateTask(task, taskCreateDto);
-        return taskService.convertEntityToDto(task);
+        TaskDto taskDto = taskService.convertEntityToDto(task);
+
+        return ResponseEntity.ok().body(taskDto);
     }
 
     @GetMapping(path = "{id}")
-    public TaskDto findTask(@PathVariable(name = "id")  Long id) {
+    public ResponseEntity<TaskDto>  findTask(@PathVariable(name = "id")  Long id) {
         Task task =  taskService.findTaskById(id);
-        return taskService.convertEntityToDto(task);
+        if(Objects.isNull(task)){
+            throw new TaskNotFoundException("Task with ID " + id + " not found");
+        }
+        TaskDto taskDto = taskService.convertEntityToDto(task);
+
+        return ResponseEntity.ok().body(taskDto);
     }
 
     @GetMapping
-    public List<TaskDto> findTasks() {
+    public ResponseEntity<List<TaskDto>> findTasks() {
         List<Task>  tasks = taskService.findTasks();
-        return taskService.convertEntitiesToDtos(tasks);
+        List<TaskDto> taskDtos = taskService.convertEntitiesToDtos(tasks);
+        return ResponseEntity.ok().body(taskDtos);
     }
 
     @DeleteMapping(path = "{id}")
     public void deleteTask(@PathVariable(name = "id")  Long id) {
         Task task =  taskService.findTaskById(id);
+        if(Objects.isNull(task)){
+            throw new TaskNotFoundException("Task with ID " + id + " not found");
+        }
         taskService.deleteTask(task);
     }
